Extract assert_parses_to into a parser utility crate
