+++ parser-utils/Cargo.toml
@@ -1,0 +1,6 @@
+[package]
+name = "parser_utils"
+version = "0.1.0"
+authors = ["Victor Porof <vporof@mozilla.com>"]
+
+[dependencies]
